Update packages to new versions with yarn upgrade --latest
