+++ package.json
@@ -6,42 +6,42 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@chakra-ui/react": "^1.8.6",
-    "@emotion/react": "^11.8.2",
+    "@chakra-ui/react": "^2.1.2",
+    "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
     "@mapbox/rehype-prism": "^0.8.0",
-    "@mdx-js/react": "^2.1.0",
+    "@mdx-js/react": "^2.1.1",
     "@mdx-js/rollup": "^2.1.1",
     "d3": "^7.4.4",
     "d3-delaunay": "^6.0.2",
-    "framer-motion": "^6.2.8",
-    "katex": "^0.15.3",
-    "plotly.js": "^2.11.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "framer-motion": "^6.3.10",
+    "katex": "^0.15.6",
+    "plotly.js": "^2.12.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-helmet": "^6.1.0",
-    "react-player": "^2.10.0",
+    "react-player": "^2.10.1",
     "react-plotly.js": "^2.5.1",
-    "react-router": "^6.2.2",
-    "react-router-dom": "^6.2.2",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
     "rehype-katex": "^6.0.2",
     "remark-gfm": "^3.0.1",
     "remark-math": "^5.1.1"
   },
   "devDependencies": {
     "@types/d3-delaunay": "^6.0.1",
-    "@types/node": "^17.0.23",
-    "@types/react": "^17.0.4",
-    "@types/react-dom": "^17.0.3",
+    "@types/node": "^17.0.39",
+    "@types/react": "^18.0.11",
+    "@types/react-dom": "^18.0.5",
     "@types/react-helmet": "^6.1.5",
-    "@types/react-plotly.js": "^2.5.0",
-    "@vitejs/plugin-react": "^1.0.7",
+    "@types/react-plotly.js": "^2.5.1",
+    "@vitejs/plugin-react": "^1.3.2",
     "chai": "^4.3.4",
-    "prettier": "^2.2.1",
+    "prettier": "^2.6.2",
     "remark-frontmatter": "^4.0.1",
     "remark-mdx-frontmatter": "^1.1.1",
-    "typescript": "^4.5.4",
-    "vite": "^2.8.0",
-    "vite-plugin-svgr": "^1.1.0"
+    "typescript": "^4.7.3",
+    "vite": "^2.9.9",
+    "vite-plugin-svgr": "^2.1.0"
   }
 }
