+++ src/pages/Projects/projects.mdx
@@ -1,5 +1,6 @@
 import { Box, Flex, Grid, Image, Center, Text, Link } from '@chakra-ui/react'
 
+import BlackHole from './blackhole'
 import Luminet from './luminet.png'
 
 ## Simulating an image of a black hole
@@ -57,6 +58,19 @@ import Luminet from './luminet.png'
                     alt="My simulated image of a black hole."
                 />
             </Link>
+        </Box>
+    </Box>
+    <Box gridColumn="span 2">
+        <Box>
+            <BlackHole />
+            <Text>
+                Here I've recreated the stippling effect used in Luminet's original figure using
+                d3.js. I used rejection sampling to randomly place stipples, and then relaxed them a
+                bit by moving them toward the centers of the voronoi cells using weights based on
+                the brightness of the original image. To make it more performant calculations are
+                done on a web worker, but it still has a terrible impact on this page's performance.
+                It's also not anywhere near as pretty as the original.
+            </Text>
         </Box>
     </Box>
     <Box gridColumn="span 3">
